+++ HISTORY.rst
@@ -3,11 +3,13 @@ Changelog
 
 0.4.0 (unreleased)
 ------------------
-- New tokenizer module with ``WordTokenizer`` and ``SentenceTokenizer``. Both textblob and NLTK tokenizer objects and be passed to TextBlob's constructor. Tokens are accessed through the new ``tokens`` property.
+- New ``text.tokenizers`` module with ``WordTokenizer`` and ``SentenceTokenizer``. Both textblob and NLTK tokenizer objects and be passed to TextBlob's constructor. Tokens are accessed through the new ``tokens`` property.
 - New ``Blobber`` class for creating TextBlobs that share the same tagger, tokenizer, and np_extractor.
 - Add ``ngrams`` method.
 - `Backwards-incompatible`: ``TextBlob.json()`` is now a method, not a property. This allows you to pass arguments (the same that you would pass to ``json.dumps()``).
 - New home for documentation: https://textblob.readthedocs.org/
+- Add parameter for cleaning HTML markup from text.
+- Minor improvement to word tokenization.
 - Fix bug with adding blobs to bytestrings.
 
 0.3.10 (2013-08-02)